[Core] ‘PlatformMemory’ is now a collection
This was a massive change-over; lots of structural changes in the layout of the file-system.

———————————————————
‘PlatformMemory’ are now ‘Collections’
———————————————————
This is the biggest change. The original point of this protocol was to have a contract that described a bag of bytes, which also had an association with a specific ‘Platform’—leaving the underlying ‘bytes’ property inaccessible publicly.

There were lots of extraneous methods being written to account or subscripting ranges on ‘PlatformMemory’, and the whole implementation was getting to feel quite messy. Now, the original goal is the reality, and the biggest benefits of this is that the way subscripts work on ‘PlatformMemory’ feel more natural.

For example, now any ‘Cartridge’ adopter has its ‘Header’ property automatically computed as an extension of the protocol itself, because instances can subscripted in such a way that doesn’t require any knowledge of the backing store.

———————————————————
Improved name spacing
———————————————————
Another monumental change is that ‘GameboyClassic’ and ‘GameboyAdvance’ derive a more reasonable namespace structure. Gone are the ‘GameboyClassicCartridge’ and ‘GameboyClassicCartridge.Header’ nonsense.

It is a much more simple: ‘GameboyClassic.Cartridge.Header.Secion’. As part of this change, some protocols lost explicit associate types AND/OR type aliases. For example, there is no ‘Cartridge’ type alias in ‘GameboyClassic’ or ‘GameboyAdvance’ anymore. The ‘Platform’ protocol still requires a ‘Cartridge’ type; however, this gets implicitly defined as part of the extension of the root platform. E.g., “GameboyClassic.Cartridge”.