[Core] improvements to protocol design
This should’ve been a lot of small commits, rather than one big one.

‘Header’ now checks for conform of a particular protocol, which allows it to be extended with all of its base properties for that given platform. It also requires that its ‘Index’ match its Platform’s ‘AddressSpace’; this ensures the ability for the Platform’s ‘HeaderSection’ to have the same ‘RawValue’ type…thus allowing the ‘HeaderSection’ to be a valid subscript index (something is also provided as part of an extension on ‘Header’.

In effect, a lot of the improvements made in these changes come from the behavior of the Platform, as was original designed but implemented in a less-than-optimal way. This also has to to do with my own familiarity with Protocol architecture in Swift improving over time as part of this project.

More interestingly, things like the ‘Cartridge.fileExtension’ can be inferred through protocol extensions now. 🎉

‘HeaderSection’ associates with a ‘Platform’, and this enables it to be used within other protocol extensions, particularly from within ‘Header’.